Implement an API for setting render targets
Each resource is created with a default state. This should be whatever the resource will be most commonly used for.

sp_graphics_command_list_set_render_targets we transition each resource from the default state into the required render target state and make a record of that transition.  We're able to assume everything is in the default state since we restore it using our records before changing render targets again or when calling close on the command list.

There's some questions with what would happen if two threads were trying to use the same resource at the same time but we' dhave problems there regardless.