Figure out triangle winding.
From gltf 2.0 specification:

For triangle primitives, the front face has a counter-clockwise (CCW) winding order.

So I set FrontCounterClockwise in the rasterizer state. That allowed me to correctly render the test model but broke others. I suspect it's because many of the materials are flagged as doubleSided. So I need to support that.