Add API for setting pipeline state.
Was originally thinking pipeline state would just be part of the material but that doesn't make sense. We need to gather state from at least the render pass and the material. Which implies we need some kind of caching and lookup that will happen at draw time. So might as well expose the function for now since it will be a higher level layer that does that.