Restore cloud shape and coverage functions.
Using a noise texture for weather instead of the debug grid. Refactored the sampling to be a little easier to read.

Also found I can force loops to not be unrolled. The compiler spits out some warnings about possibly undefined partial derivatives but I'm not seeing any problem in the sampling. And it vastly improves compile times. Worth it for now.